ref(performance): Split project performance settings
Move detector configuration and detector form fields into focused modules while preserving the existing page API and behavior.
